navbar, footer and basic routes
