+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{  
+  "workbench.colorCustomizations": {
+    "titleBar.activeForeground": "#000000",
+    "titleBar.inactiveForeground": "#000000CC",
+    "titleBar.activeBackground": "#68b9e6",
+    "titleBar.inactiveBackground": "#307397"
+  }
+}
