+++ src/lib/footer.svelte
@@ -1,0 +1,5 @@
+	<footer class="bg-gray-800 px-8 py-12 text-gray-100">
+		<div class="mx-auto max-w-7xl">
+			&copy; {new Date().getFullYear()} Accelerated Equity Plans
+		</div>
+	</footer>
