+++ src/routes/about/+page.svelte
@@ -1,0 +1,4 @@
+<div>
+  <h1 class="text-2xl">This is the about page</h1>
+  <p>About the people that made this site</p>
+</div>
