+++ src/routes/careers/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Careers Page</h1>
