+++ src/routes/contact/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Contact Page</h1>
