+++ src/routes/services/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Our Services</h1>
