+++ src/routes/testimonials/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Testimonials about how awesome we are!</h1>
